Revert "feat: mobile wallet new flow (#8689)"
This reverts commit 7a87bf4e78263325e0ba4a622c21379763a78b95.
