+++ package.json
@@ -208,7 +208,7 @@
     "stream-http": "^3.2.0",
     "styled-components": "^6.0.7",
     "uuid": "^9.0.0",
-    "vaul": "^1.1.2",
+    "vaul": "^0.9.0",
     "viem": "^2.10.9",
     "wagmi": "^2.9.2",
     "web-vitals": "^2.1.4"
