+++ src/assets/translations/en/main.json
@@ -11,12 +11,10 @@
     "and": "and",
     "balance": "Balance",
     "next": "Next",
-    "edit": "Edit",
     "error": "Error",
     "show": "Show",
     "account": "Account",
     "address": "Address",
-    "noThanks": "No Thanks",
     "done": "Done",
     "cancel": "Cancel",
     "close": "Close",
@@ -1240,11 +1238,6 @@
         "menuItem": "Back up my wallet",
         "enterPassword": "Enter your password to continue",
         "testTitle": "Verify Your Secret Recovery Phrase",
-        "confirm": {
-          "title": "Backup wallet",
-          "body": "Before you forget your wallet, would you like to back up your secret recovery phrase?",
-          "backupNow": "Backup Now"
-        },
         "info": {
           "title": "Your Secret Recovery Phrase",
           "description": "Write these 12 words down and store them securely offline. This 12 word phrase is used to recover your wallet private keys.",
@@ -1771,15 +1764,12 @@
         "error": {
           "delete": "Unable to delete your wallet, sorry.",
           "noWallet": "You have no saved wallets",
-          "pair": "Unable to pair your wallet",
-          "fetchingWallets": "An error occurred while fetching wallets."
+          "pair": "Unable to pair your wallet"
         },
         "header": "Saved Wallets",
         "body": "Loading your saved wallet... Enter your password when prompted.",
         "button": "Continue",
-        "forgetWallet": "Forget Wallet",
-        "confirmForget": "Are you sure you want to forget %{wallet}?",
-        "confirmForgetBody": "You will not be able to access your wallet if you have not backed it up."
+        "confirmForget": "Are you sure you want to forget %{wallet}?"
       },
       "import": {
         "header": "Import your wallet",
