--- src/components/MobileWalletDialog/types.ts
@@ -1,7 +1,0 @@
-export enum MobileWalletDialogRoutes {
-  Saved = '/saved',
-  Rename = '/rename',
-  Delete = '/delete',
-  Backup = '/delete/backup',
-  ConfirmDelete = '/delete/confirm',
-}
