+++ src/context/WalletProvider/MobileWallet/components/MobileCreateTest.tsx
@@ -1,5 +1,4 @@
 import { Button, ModalBody, ModalHeader, SimpleGrid, Tag } from '@chakra-ui/react'
-import { useQueryClient } from '@tanstack/react-query'
 import * as bip39 from 'bip39'
 import range from 'lodash/range'
 import shuffle from 'lodash/shuffle'
@@ -59,7 +58,6 @@ export const MobileCreateTest = ({ history, location }: MobileSetupProps) => {
   const [testCount, setTestCount] = useState<number>(0)
   const shuffledNumbers = useMemo(() => slice(shuffle(range(12)), 0, TEST_COUNT_REQUIRED), [])
   const revoker = useMemo(() => new (Revocable(class {}))(), [])
-  const queryClient = useQueryClient()
 
   const { vault } = location.state
 
@@ -101,8 +99,6 @@ export const MobileCreateTest = ({ history, location }: MobileSetupProps) => {
       ;(async () => {
         if (vault?.label && vault?.mnemonic) {
           const newWallet = await addWallet({ label: vault.label, mnemonic: vault.mnemonic })
-
-          queryClient.invalidateQueries({ queryKey: ['listWallets'] })
           history.replace('/mobile/success', { vault: newWallet! })
         }
       })()
@@ -112,7 +108,7 @@ export const MobileCreateTest = ({ history, location }: MobileSetupProps) => {
         setTimeout(() => revoker.revoke(), 250)
       }
     }
-  }, [testCount, history, vault, revoker, queryClient])
+  }, [testCount, history, vault, revoker])
 
   const handleClick = useCallback(
     (index: number) => {
