+++ src/context/WalletProvider/MobileWallet/components/MobileImport.tsx
@@ -7,7 +7,6 @@ import {
   ModalHeader,
   Textarea,
 } from '@chakra-ui/react'
-import { useQueryClient } from '@tanstack/react-query'
 import * as bip39 from 'bip39'
 import { useCallback, useMemo } from 'react'
 import { useForm } from 'react-hook-form'
@@ -26,7 +25,6 @@ export const MobileImport = ({ history }: RouteComponentProps) => {
     register,
     formState: { errors, isSubmitting },
   } = useForm<FormValues>({ shouldUnregister: true })
-  const queryClient = useQueryClient()
 
   const onSubmit = useCallback(
     async (values: FormValues) => {
@@ -37,13 +35,12 @@ export const MobileImport = ({ history }: RouteComponentProps) => {
           label: values.name.trim(),
         })
         history.push('/mobile/success', { vault })
-        queryClient.invalidateQueries({ queryKey: ['listWallets'] })
       } catch (e) {
         console.log(e)
         setError('mnemonic', { type: 'manual', message: 'walletProvider.shapeShift.import.header' })
       }
     },
-    [history, queryClient, setError],
+    [history, setError],
   )
 
   const translate = useTranslate()
