+++ src/pages/Dashboard/components/DashboardHeader/DashboardHeaderTop.tsx
@@ -3,15 +3,12 @@ import { Button, Container, Flex, IconButton, useDisclosure } from '@chakra-ui/r
 import type { ResponsiveValue } from '@chakra-ui/system'
 import type { Property } from 'csstype'
 import { memo, useCallback } from 'react'
-import { FiLogOut } from 'react-icons/fi'
 import { IoEllipsisHorizontal, IoSwapVerticalSharp } from 'react-icons/io5'
 import { useTranslate } from 'react-polyglot'
 import { QRCodeIcon } from 'components/Icons/QRCode'
-import { MobileWalletDialog } from 'components/MobileWalletDialog/MobileWalletDialog'
 import { useBrowserRouter } from 'hooks/useBrowserRouter/useBrowserRouter'
 import { useModal } from 'hooks/useModal/useModal'
 import { useWallet } from 'hooks/useWallet/useWallet'
-import { isMobile } from 'lib/globals'
 
 import { EditAvatarButton, ProfileAvatar } from '../ProfileAvatar/ProfileAvatar'
 import { DashboardDrawer } from './DashboardDrawer'
@@ -21,7 +18,7 @@ const qrCodeIcon = <QRCodeIcon />
 const arrowUpIcon = <ArrowUpIcon />
 const arrowDownIcon = <ArrowDownIcon />
 const ioSwapVerticalSharpIcon = <IoSwapVerticalSharp />
-const moreIcon = isMobile ? <FiLogOut /> : <IoEllipsisHorizontal />
+const moreIcon = <IoEllipsisHorizontal />
 
 const ButtonRowDisplay = { base: 'flex', md: 'none' }
 
@@ -113,11 +110,7 @@ export const DashboardHeaderTop = memo(() => {
       >
         <IconButton isRound icon={moreIcon} aria-label='Settings' onClick={onOpen} />
       </Flex>
-      {isMobile ? (
-        <MobileWalletDialog isOpen={isOpen} onClose={onClose} />
-      ) : (
-        <DashboardDrawer isOpen={isOpen} onClose={onClose} />
-      )}
+      <DashboardDrawer isOpen={isOpen} onClose={onClose} />
     </Container>
   )
 })
